+++ layrz_sdk/entities/device.py
@@ -2,6 +2,8 @@
 
 from pydantic import BaseModel, Field
 
+from .modbus import ModbusConfig
+
 
 class Device(BaseModel):
   """Device entity"""
@@ -9,5 +11,8 @@ class Device(BaseModel):
   pk: int = Field(description='Defines the primary key of the device')
   name: str = Field(description='Defines the name of the device')
   ident: str = Field(description='Defines the identifier of the device')
+  protocol_id: int = Field(description='Defines the protocol ID of the device')
   protocol: str = Field(description='Defines the protocol of the device')
   is_primary: bool = Field(default=False, description='Defines if the device is the primary device')
+
+  modbus: ModbusConfig | None = Field(default=None, description='Modbus configuration')